+++ lib/PipeTransport.d.ts
@@ -2,7 +2,8 @@ import { EnhancedEventEmitter } from './EnhancedEventEmitter';
 import { Transport, TransportListenIp, TransportTuple, SctpState } from './Transport';
 import { Consumer, ConsumerOptions } from './Consumer';
 import { SctpParameters, NumSctpStreams } from './SctpParameters';
-export interface PipeTransportOptions {
+import { SrtpParameters } from './SrtpParameters';
+export declare type PipeTransportOptions = {
     /**
      * Listening IP address.
      */
@@ -21,11 +22,23 @@ export interface PipeTransportOptions {
      */
     maxSctpMessageSize?: number;
     /**
+     * Enable RTX and NACK for RTP retransmission. Useful if both Routers are
+     * located in different hosts and there is packet lost in the link. For this
+     * to work, both PipeTransports must enable this setting. Default false.
+     */
+    enableRtx?: boolean;
+    /**
+     * Enable SRTP. Useful to protect the RTP and RTCP traffic if both Routers
+     * are located in different hosts. For this to work, connect() must be called
+     * with remote SRTP parameters. Default false.
+     */
+    enableSrtp?: boolean;
+    /**
      * Custom application data.
      */
     appData?: any;
-}
-export interface PipeTransportStat {
+};
+export declare type PipeTransportStat = {
     type: string;
     transportId: string;
     timestamp: number;
@@ -50,7 +63,7 @@ export interface PipeTransportStat {
     availableIncomingBitrate?: number;
     maxIncomingBitrate?: number;
     tuple: TransportTuple;
-}
+};
 export declare class PipeTransport extends Transport {
     /**
      * @private
@@ -71,6 +84,10 @@ export declare class PipeTransport extends Transport {
      */
     get sctpState(): SctpState;
     /**
+     * SRTP parameters.
+     */
+    get srtpParameters(): SrtpParameters | undefined;
+    /**
      * Observer.
      *
      * @override
@@ -84,7 +101,7 @@ export declare class PipeTransport extends Transport {
      */
     get observer(): EnhancedEventEmitter;
     /**
-     * Close the PlainRtpTransport.
+     * Close the PipeTransport.
      *
      * @override
      */
@@ -107,9 +124,10 @@ export declare class PipeTransport extends Transport {
      *
      * @override
      */
-    connect({ ip, port }: {
+    connect({ ip, port, srtpParameters }: {
         ip: string;
         port: number;
+        srtpParameters?: SrtpParameters;
     }): Promise<void>;
     /**
      * Create a pipe Consumer.