+++ lib/PlainTransport.d.ts
@@ -2,7 +2,8 @@ import { EnhancedEventEmitter } from './EnhancedEventEmitter';
 import { Transport, TransportListenIp, TransportTuple, SctpState } from './Transport';
 import { Consumer, ConsumerOptions } from './Consumer';
 import { SctpParameters, NumSctpStreams } from './SctpParameters';
-export interface PlainRtpTransportOptions {
+import { SrtpParameters, SrtpCryptoSuite } from './SrtpParameters';
+export declare type PlainTransportOptions = {
     /**
      * Listening IP address.
      */
@@ -37,11 +38,25 @@ export interface PlainRtpTransportOptions {
      */
     maxSctpMessageSize?: number;
     /**
+     * Enable SRTP. For this to work, connect() must be called
+     * with remote SRTP parameters. Default false.
+     */
+    enableSrtp?: boolean;
+    /**
+     * The SRTP crypto suite to be used if enableSrtp is set. Default
+     * 'AES_CM_128_HMAC_SHA1_80'.
+     */
+    srtpCryptoSuite?: SrtpCryptoSuite;
+    /**
      * Custom application data.
      */
     appData?: any;
-}
-export interface PlainRtpTransportStat {
+};
+/**
+ * DEPRECATED: Use PlainTransportOptions.
+ */
+export declare type PlainRtpTransportOptions = PlainTransportOptions;
+export declare type PlainTransportStat = {
     type: string;
     transportId: string;
     timestamp: number;
@@ -70,10 +85,16 @@ export interface PlainRtpTransportStat {
     multiSource: boolean;
     tuple: TransportTuple;
     rtcpTuple?: TransportTuple;
-}
-export declare class PlainRtpTransport extends Transport {
+};
+/**
+ * DEPRECATED: Use PlainTransportOptions.
+ */
+export declare type PlainRtpTransportStat = PlainTransportStat;
+export declare class PlainTransport extends Transport {
     /**
      * @private
+     * @emits tuple - (tuple: TransportTuple)
+     * @emits rtcpTuple - (rtcpTuple: TransportTuple)
      * @emits sctpstatechange - (sctpState: SctpState)
      * @emits trace - (trace: TransportTraceEventData)
      */
@@ -85,7 +106,7 @@ export declare class PlainRtpTransport extends Transport {
     /**
      * Transport RTCP tuple.
      */
-    get rtcpTuple(): TransportTuple;
+    get rtcpTuple(): TransportTuple | undefined;
     /**
      * SCTP parameters.
      */
@@ -95,6 +116,10 @@ export declare class PlainRtpTransport extends Transport {
      */
     get sctpState(): SctpState;
     /**
+     * SRTP parameters.
+     */
+    get srtpParameters(): SrtpParameters | undefined;
+    /**
      * Observer.
      *
      * @override
@@ -103,12 +128,14 @@ export declare class PlainRtpTransport extends Transport {
      * @emits newconsumer - (producer: Producer)
      * @emits newdataproducer - (dataProducer: DataProducer)
      * @emits newdataconsumer - (dataProducer: DataProducer)
+     * @emits tuple - (tuple: TransportTuple)
+     * @emits rtcpTuple - (rtcpTuple: TransportTuple)
      * @emits sctpstatechange - (sctpState: SctpState)
      * @emits trace - (trace: TransportTraceEventData)
      */
     get observer(): EnhancedEventEmitter;
     /**
-     * Close the PlainRtpTransport.
+     * Close the PlainTransport.
      *
      * @override
      */
@@ -121,20 +148,21 @@ export declare class PlainRtpTransport extends Transport {
      */
     routerClosed(): void;
     /**
-     * Get PlainRtpTransport stats.
+     * Get PlainTransport stats.
      *
      * @override
      */
-    getStats(): Promise<PlainRtpTransportStat[]>;
+    getStats(): Promise<PlainTransportStat[]>;
     /**
-     * Provide the PlainRtpTransport remote parameters.
+     * Provide the PlainTransport remote parameters.
      *
      * @override
      */
-    connect({ ip, port, rtcpPort }: {
-        ip: string;
-        port: number;
+    connect({ ip, port, rtcpPort, srtpParameters }: {
+        ip?: string;
+        port?: number;
         rtcpPort?: number;
+        srtpParameters?: SrtpParameters;
     }): Promise<void>;
     /**
      * Override Transport.consume() method to reject it if multiSource is set.
@@ -144,4 +172,10 @@ export declare class PlainRtpTransport extends Transport {
     consume(params: ConsumerOptions): Promise<Consumer>;
     private _handleWorkerNotifications;
 }
-//# sourceMappingURL=PlainRtpTransport.d.ts.map
+/**
+ * DEPRECATED: Use PlainTransport.
+ */
+export declare class PlainRtpTransport extends PlainTransport {
+    constructor(params: any);
+}
+//# sourceMappingURL=PlainTransport.d.ts.map