+++ worker/src/RTC/PlainTransport.cpp
@@ -1,18 +1,36 @@
-#define MS_CLASS "RTC::PlainRtpTransport"
+#define MS_CLASS "RTC::PlainTransport"
 // #define MS_LOG_DEV_LEVEL 3
 
-#include "RTC/PlainRtpTransport.hpp"
+#include "RTC/PlainTransport.hpp"
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
 #include "Utils.hpp"
+#include "Channel/Notifier.hpp"
 
 namespace RTC
 {
+	/* Static. */
+
+	// NOTE: PlainTransport allows AES_CM_128_HMAC_SHA1_80 and
+	// AES_CM_128_HMAC_SHA1_32 SRTP crypto suites.
+	// clang-format off
+	std::map<std::string, RTC::SrtpSession::CryptoSuite> PlainTransport::string2SrtpCryptoSuite =
+	{
+		{ "AES_CM_128_HMAC_SHA1_80", RTC::SrtpSession::CryptoSuite::AES_CM_128_HMAC_SHA1_80 },
+		{ "AES_CM_128_HMAC_SHA1_32", RTC::SrtpSession::CryptoSuite::AES_CM_128_HMAC_SHA1_32 }
+	};
+	std::map<RTC::SrtpSession::CryptoSuite, std::string> PlainTransport::srtpCryptoSuite2String =
+	{
+		{ RTC::SrtpSession::CryptoSuite::AES_CM_128_HMAC_SHA1_80, "AES_CM_128_HMAC_SHA1_80" },
+		{ RTC::SrtpSession::CryptoSuite::AES_CM_128_HMAC_SHA1_32, "AES_CM_128_HMAC_SHA1_32" }
+	};
+	// clang-format on
+	size_t PlainTransport::srtpMasterLength{ 30 };
+
 	/* Instance methods. */
 
 	// NOLINTNEXTLINE(cppcoreguidelines-pro-type-member-init)
-	PlainRtpTransport::PlainRtpTransport(
-	  const std::string& id, RTC::Transport::Listener* listener, json& data)
+	PlainTransport::PlainTransport(const std::string& id, RTC::Transport::Listener* listener, json& data)
 	  : RTC::Transport::Transport(id, listener, data)
 	{
 		MS_TRACE();
@@ -83,6 +101,34 @@ namespace RTC
 			}
 		}
 
+		auto jsonEnableSrtpIt = data.find("enableSrtp");
+
+		// clang-format off
+		if (
+			jsonEnableSrtpIt != data.end() &&
+			jsonEnableSrtpIt->is_boolean() &&
+			jsonEnableSrtpIt->get<bool>()
+		)
+		// clang-format on
+		{
+			auto jsonSrtpCryptoSuiteIt = data.find("srtpCryptoSuite");
+
+			if (jsonSrtpCryptoSuiteIt == data.end() || !jsonSrtpCryptoSuiteIt->is_string())
+				MS_THROW_TYPE_ERROR("missing srtpCryptoSuite)");
+
+			// Ensure it's a crypto suite supported by us.
+			auto it =
+			  PlainTransport::string2SrtpCryptoSuite.find(jsonSrtpCryptoSuiteIt->get<std::string>());
+
+			if (it == PlainTransport::string2SrtpCryptoSuite.end())
+				MS_THROW_TYPE_ERROR("invalid/unsupported srtpCryptoSuite");
+
+			// NOTE: The SRTP crypto suite may change later on connect().
+			this->srtpCryptoSuite = it->second;
+			this->srtpKey         = Utils::Crypto::GetRandomString(PlainTransport::srtpMasterLength);
+			this->srtpKeyBase64   = Utils::String::Base64Encode(this->srtpKey);
+		}
+
 		try
 		{
 			// This may throw.
@@ -106,7 +152,7 @@ namespace RTC
 		}
 	}
 
-	PlainRtpTransport::~PlainRtpTransport()
+	PlainTransport::~PlainTransport()
 	{
 		MS_TRACE();
 
@@ -121,9 +167,15 @@ namespace RTC
 
 		delete this->rtcpTuple;
 		this->rtcpTuple = nullptr;
+
+		delete this->srtpSendSession;
+		this->srtpSendSession = nullptr;
+
+		delete this->srtpRecvSession;
+		this->srtpRecvSession = nullptr;
 	}
 
-	void PlainRtpTransport::FillJson(json& jsonObject) const
+	void PlainTransport::FillJson(json& jsonObject) const
 	{
 		MS_TRACE();
 
@@ -140,7 +192,7 @@ namespace RTC
 		jsonObject["multiSource"] = this->multiSource;
 
 		// Add tuple.
-		if (this->tuple != nullptr)
+		if (this->tuple)
 		{
 			this->tuple->FillJson(jsonObject["tuple"]);
 		}
@@ -161,7 +213,7 @@ namespace RTC
 		// Add rtcpTuple.
 		if (!this->rtcpMux)
 		{
-			if (this->rtcpTuple != nullptr)
+			if (this->rtcpTuple)
 			{
 				this->rtcpTuple->FillJson(jsonObject["rtcpTuple"]);
 			}
@@ -179,9 +231,20 @@ namespace RTC
 				(*jsonRtcpTupleIt)["protocol"]  = "udp";
 			}
 		}
+
+		// Add srtpParameters.
+		if (HasSrtp())
+		{
+			jsonObject["srtpParameters"] = json::object();
+			auto jsonSrtpParametersIt    = jsonObject.find("srtpParameters");
+
+			(*jsonSrtpParametersIt)["cryptoSuite"] =
+			  PlainTransport::srtpCryptoSuite2String[this->srtpCryptoSuite];
+			(*jsonSrtpParametersIt)["keyBase64"] = this->srtpKeyBase64;
+		}
 	}
 
-	void PlainRtpTransport::FillJsonStats(json& jsonArray)
+	void PlainTransport::FillJsonStats(json& jsonArray)
 	{
 		MS_TRACE();
 
@@ -202,7 +265,7 @@ namespace RTC
 		// Add multiSource.
 		jsonObject["multiSource"] = this->multiSource;
 
-		if (this->tuple != nullptr)
+		if (this->tuple)
 		{
 			// Add tuple.
 			this->tuple->FillJson(jsonObject["tuple"]);
@@ -223,11 +286,11 @@ namespace RTC
 		}
 
 		// Add rtcpTuple.
-		if (!this->rtcpMux && this->rtcpTuple != nullptr)
+		if (!this->rtcpMux && this->rtcpTuple)
 			this->rtcpTuple->FillJson(jsonObject["rtcpTuple"]);
 	}
 
-	void PlainRtpTransport::HandleRequest(Channel::Request* request)
+	void PlainTransport::HandleRequest(Channel::Request* request)
 	{
 		MS_TRACE();
 
@@ -235,14 +298,8 @@ namespace RTC
 		{
 			case Channel::Request::MethodId::TRANSPORT_CONNECT:
 			{
-				// Reject if comedia mode or multiSource is set.
-				if (this->comedia)
-					MS_THROW_ERROR("cannot call connect() when comedia mode is set");
-				else if (this->multiSource)
-					MS_THROW_ERROR("cannot call connect() when multiSource is set");
-
 				// Ensure this method is not called twice.
-				if (this->tuple != nullptr)
+				if (this->connectCalled)
 					MS_THROW_ERROR("connect() already called");
 
 				try
@@ -250,107 +307,170 @@ namespace RTC
 					std::string ip;
 					uint16_t port{ 0u };
 					uint16_t rtcpPort{ 0u };
+					std::string srtpKeyBase64;
 
-					auto jsonIpIt = request->data.find("ip");
+					auto jsonSrtpParametersIt = request->data.find("srtpParameters");
 
-					if (jsonIpIt == request->data.end() || !jsonIpIt->is_string())
-						MS_THROW_TYPE_ERROR("missing ip");
-
-					ip = jsonIpIt->get<std::string>();
-
-					// This may throw.
-					Utils::IP::NormalizeIp(ip);
-
-					auto jsonPortIt = request->data.find("port");
-
-					// clang-format off
-					if (
-						jsonPortIt == request->data.end() ||
-						!Utils::Json::IsPositiveInteger(*jsonPortIt)
-					)
-					// clang-format on
+					if (!HasSrtp() && jsonSrtpParametersIt != request->data.end())
 					{
-						MS_THROW_TYPE_ERROR("missing port");
+						MS_THROW_TYPE_ERROR("invalid srtpParameters (SRTP not enabled)");
 					}
-
-					port = jsonPortIt->get<uint16_t>();
-
-					auto jsonRtcpPortIt = request->data.find("rtcpPort");
-
-					// clang-format off
-					if (
-						jsonRtcpPortIt != request->data.end() &&
-						Utils::Json::IsPositiveInteger(*jsonRtcpPortIt)
-					)
-					// clang-format on
+					else if (HasSrtp())
 					{
-						if (this->rtcpMux)
-							MS_THROW_TYPE_ERROR("cannot set rtcpPort with rtcpMux enabled");
-
-						rtcpPort = jsonRtcpPortIt->get<uint16_t>();
-					}
-					else
-					{
-						if (!this->rtcpMux)
-							MS_THROW_TYPE_ERROR("missing rtcpPort (required with rtcpMux disabled)");
-					}
-
-					int err;
-
-					switch (Utils::IP::GetFamily(ip))
-					{
-						case AF_INET:
+						// clang-format off
+						if (
+							jsonSrtpParametersIt == request->data.end() ||
+							!jsonSrtpParametersIt->is_object()
+						)
+						// clang-format on
 						{
-							err = uv_ip4_addr(
-							  ip.c_str(),
-							  static_cast<int>(port),
-							  reinterpret_cast<struct sockaddr_in*>(&this->remoteAddrStorage));
-
-							if (err != 0)
-								MS_ABORT("uv_ip4_addr() failed: %s", uv_strerror(err));
-
-							break;
+							MS_THROW_TYPE_ERROR("missing srtpParameters (SRTP enabled)");
 						}
 
-						case AF_INET6:
+						auto jsonCryptoSuiteIt = jsonSrtpParametersIt->find("cryptoSuite");
+
+						// clang-format off
+						if (
+							jsonCryptoSuiteIt == jsonSrtpParametersIt->end() ||
+							!jsonCryptoSuiteIt->is_string()
+						)
+						// clang-format on
 						{
-							err = uv_ip6_addr(
-							  ip.c_str(),
-							  static_cast<int>(port),
-							  reinterpret_cast<struct sockaddr_in6*>(&this->remoteAddrStorage));
-
-							if (err != 0)
-								MS_ABORT("uv_ip6_addr() failed: %s", uv_strerror(err));
-
-							break;
+							MS_THROW_TYPE_ERROR("missing srtpParameters.cryptoSuite)");
 						}
 
-						default:
+						// Ensure it's a crypto suite supported by us.
+						auto it =
+						  PlainTransport::string2SrtpCryptoSuite.find(jsonCryptoSuiteIt->get<std::string>());
+
+						if (it == PlainTransport::string2SrtpCryptoSuite.end())
+							MS_THROW_TYPE_ERROR("invalid/unsupported srtpParameters.cryptoSuite");
+
+						// Update out SRTP crypto suite wuth the one used by the remote.
+						this->srtpCryptoSuite = it->second;
+
+						auto jsonKeyBase64It = jsonSrtpParametersIt->find("keyBase64");
+
+						// clang-format off
+						if (
+							jsonKeyBase64It == jsonSrtpParametersIt->end() ||
+							!jsonKeyBase64It->is_string()
+						)
+						// clang-format on
 						{
-							MS_THROW_ERROR("invalid IP '%s'", ip.c_str());
+							MS_THROW_TYPE_ERROR("missing srtpParameters.keyBase64)");
 						}
+
+						srtpKeyBase64 = jsonKeyBase64It->get<std::string>();
+
+						size_t outLen;
+						// This may throw.
+						auto* srtpKey = Utils::String::Base64Decode(srtpKeyBase64, outLen);
+
+						if (outLen != PlainTransport::srtpMasterLength)
+							MS_THROW_TYPE_ERROR("invalid decoded SRTP key length");
+
+						auto* srtpLocalKey  = new uint8_t[PlainTransport::srtpMasterLength];
+						auto* srtpRemoteKey = new uint8_t[PlainTransport::srtpMasterLength];
+
+						std::memcpy(srtpLocalKey, this->srtpKey.c_str(), PlainTransport::srtpMasterLength);
+						std::memcpy(srtpRemoteKey, srtpKey, PlainTransport::srtpMasterLength);
+
+						try
+						{
+							this->srtpSendSession = new RTC::SrtpSession(
+							  RTC::SrtpSession::Type::OUTBOUND,
+							  this->srtpCryptoSuite,
+							  srtpLocalKey,
+							  PlainTransport::srtpMasterLength);
+						}
+						catch (const MediaSoupError& error)
+						{
+							delete[] srtpLocalKey;
+							delete[] srtpRemoteKey;
+
+							MS_THROW_ERROR("error creating SRTP sending session: %s", error.what());
+						}
+
+						try
+						{
+							this->srtpRecvSession = new RTC::SrtpSession(
+							  RTC::SrtpSession::Type::INBOUND,
+							  this->srtpCryptoSuite,
+							  srtpRemoteKey,
+							  PlainTransport::srtpMasterLength);
+						}
+						catch (const MediaSoupError& error)
+						{
+							delete[] srtpLocalKey;
+							delete[] srtpRemoteKey;
+
+							MS_THROW_ERROR("error creating SRTP receiving session: %s", error.what());
+						}
+
+						delete[] srtpLocalKey;
+						delete[] srtpRemoteKey;
 					}
 
-					// Create the tuple.
-					this->tuple = new RTC::TransportTuple(
-					  this->udpSocket, reinterpret_cast<struct sockaddr*>(&this->remoteAddrStorage));
-
-					if (!this->listenIp.announcedIp.empty())
-						this->tuple->SetLocalAnnouncedIp(this->listenIp.announcedIp);
-
-					if (!this->rtcpMux)
+					if (!this->comedia && !this->multiSource)
 					{
+						auto jsonIpIt = request->data.find("ip");
+
+						if (jsonIpIt == request->data.end() || !jsonIpIt->is_string())
+							MS_THROW_TYPE_ERROR("missing ip");
+
+						ip = jsonIpIt->get<std::string>();
+
+						// This may throw.
+						Utils::IP::NormalizeIp(ip);
+
+						auto jsonPortIt = request->data.find("port");
+
+						// clang-format off
+						if (
+							jsonPortIt == request->data.end() ||
+							!Utils::Json::IsPositiveInteger(*jsonPortIt)
+						)
+						// clang-format on
+						{
+							MS_THROW_TYPE_ERROR("missing port");
+						}
+
+						port = jsonPortIt->get<uint16_t>();
+
+						auto jsonRtcpPortIt = request->data.find("rtcpPort");
+
+						// clang-format off
+						if (
+							jsonRtcpPortIt != request->data.end() &&
+							Utils::Json::IsPositiveInteger(*jsonRtcpPortIt)
+						)
+						// clang-format on
+						{
+							if (this->rtcpMux)
+								MS_THROW_TYPE_ERROR("cannot set rtcpPort with rtcpMux enabled");
+
+							rtcpPort = jsonRtcpPortIt->get<uint16_t>();
+						}
+						else
+						{
+							if (!this->rtcpMux)
+								MS_THROW_TYPE_ERROR("missing rtcpPort (required with rtcpMux disabled)");
+						}
+
+						int err;
+
 						switch (Utils::IP::GetFamily(ip))
 						{
 							case AF_INET:
 							{
 								err = uv_ip4_addr(
 								  ip.c_str(),
-								  static_cast<int>(rtcpPort),
-								  reinterpret_cast<struct sockaddr_in*>(&this->rtcpRemoteAddrStorage));
+								  static_cast<int>(port),
+								  reinterpret_cast<struct sockaddr_in*>(&this->remoteAddrStorage));
 
 								if (err != 0)
-									MS_ABORT("uv_ip4_addr() failed: %s", uv_strerror(err));
+									MS_THROW_ERROR("uv_ip4_addr() failed: %s", uv_strerror(err));
 
 								break;
 							}
@@ -359,11 +479,11 @@ namespace RTC
 							{
 								err = uv_ip6_addr(
 								  ip.c_str(),
-								  static_cast<int>(rtcpPort),
-								  reinterpret_cast<struct sockaddr_in6*>(&this->rtcpRemoteAddrStorage));
+								  static_cast<int>(port),
+								  reinterpret_cast<struct sockaddr_in6*>(&this->remoteAddrStorage));
 
 								if (err != 0)
-									MS_ABORT("uv_ip6_addr() failed: %s", uv_strerror(err));
+									MS_THROW_ERROR("uv_ip6_addr() failed: %s", uv_strerror(err));
 
 								break;
 							}
@@ -375,37 +495,95 @@ namespace RTC
 						}
 
 						// Create the tuple.
-						this->rtcpTuple = new RTC::TransportTuple(
-						  this->rtcpUdpSocket, reinterpret_cast<struct sockaddr*>(&this->rtcpRemoteAddrStorage));
+						this->tuple = new RTC::TransportTuple(
+						  this->udpSocket, reinterpret_cast<struct sockaddr*>(&this->remoteAddrStorage));
 
 						if (!this->listenIp.announcedIp.empty())
-							this->rtcpTuple->SetLocalAnnouncedIp(this->listenIp.announcedIp);
+							this->tuple->SetLocalAnnouncedIp(this->listenIp.announcedIp);
+
+						if (!this->rtcpMux)
+						{
+							switch (Utils::IP::GetFamily(ip))
+							{
+								case AF_INET:
+								{
+									err = uv_ip4_addr(
+									  ip.c_str(),
+									  static_cast<int>(rtcpPort),
+									  reinterpret_cast<struct sockaddr_in*>(&this->rtcpRemoteAddrStorage));
+
+									if (err != 0)
+										MS_THROW_ERROR("uv_ip4_addr() failed: %s", uv_strerror(err));
+
+									break;
+								}
+
+								case AF_INET6:
+								{
+									err = uv_ip6_addr(
+									  ip.c_str(),
+									  static_cast<int>(rtcpPort),
+									  reinterpret_cast<struct sockaddr_in6*>(&this->rtcpRemoteAddrStorage));
+
+									if (err != 0)
+										MS_THROW_ERROR("uv_ip6_addr() failed: %s", uv_strerror(err));
+
+									break;
+								}
+
+								default:
+								{
+									MS_THROW_ERROR("invalid IP '%s'", ip.c_str());
+								}
+							}
+
+							// Create the tuple.
+							this->rtcpTuple = new RTC::TransportTuple(
+							  this->rtcpUdpSocket,
+							  reinterpret_cast<struct sockaddr*>(&this->rtcpRemoteAddrStorage));
+
+							if (!this->listenIp.announcedIp.empty())
+								this->rtcpTuple->SetLocalAnnouncedIp(this->listenIp.announcedIp);
+						}
 					}
 				}
 				catch (const MediaSoupError& error)
 				{
-					if (this->tuple != nullptr)
-					{
-						delete this->tuple;
-						this->tuple = nullptr;
-					}
+					delete this->tuple;
+					this->tuple = nullptr;
 
-					if (this->rtcpTuple != nullptr)
-					{
-						delete this->rtcpTuple;
-						this->rtcpTuple = nullptr;
-					}
+					delete this->rtcpTuple;
+					this->rtcpTuple = nullptr;
+
+					delete this->srtpSendSession;
+					this->srtpSendSession = nullptr;
+
+					delete this->srtpRecvSession;
+					this->srtpRecvSession = nullptr;
 
 					throw;
 				}
 
+				this->connectCalled = true;
+
 				// Tell the caller about the selected local DTLS role.
 				json data = json::object();
 
-				this->tuple->FillJson(data["tuple"]);
+				if (this->tuple)
+					this->tuple->FillJson(data["tuple"]);
 
-				if (!this->rtcpMux)
+				if (!this->rtcpMux && this->rtcpTuple)
 					this->rtcpTuple->FillJson(data["rtcpTuple"]);
+
+				if (HasSrtp())
+				{
+					data["srtpParameters"]    = json::object();
+					auto jsonSrtpParametersIt = data.find("srtpParameters");
+
+					(*jsonSrtpParametersIt)["cryptoSuite"] =
+					  PlainTransport::srtpCryptoSuite2String[this->srtpCryptoSuite];
+					(*jsonSrtpParametersIt)["keyBase64"] = this->srtpKeyBase64;
+				}
 
 				request->Accept(data);
 
@@ -424,12 +602,22 @@ namespace RTC
 		}
 	}
 
-	inline bool PlainRtpTransport::IsConnected() const
+	inline bool PlainTransport::IsConnected() const
 	{
-		return this->tuple != nullptr;
+		return this->tuple;
 	}
 
-	void PlainRtpTransport::SendRtpPacket(RTC::RtpPacket* packet, RTC::Transport::onSendCallback* cb)
+	inline bool PlainTransport::HasSrtp() const
+	{
+		return !this->srtpKey.empty();
+	}
+
+	inline bool PlainTransport::IsSrtpReady() const
+	{
+		return HasSrtp() && this->srtpSendSession && this->srtpRecvSession;
+	}
+
+	void PlainTransport::SendRtpPacket(RTC::RtpPacket* packet, RTC::Transport::onSendCallback* cb)
 	{
 		MS_TRACE();
 
@@ -448,13 +636,25 @@ namespace RTC
 		const uint8_t* data = packet->GetData();
 		size_t len          = packet->GetSize();
 
+		if (HasSrtp() && !this->srtpSendSession->EncryptRtp(&data, &len))
+		{
+			if (cb)
+			{
+				(*cb)(false);
+
+				delete cb;
+			}
+
+			return;
+		}
+
 		this->tuple->Send(data, len, cb);
 
 		// Increase send transmission.
 		RTC::Transport::DataSent(len);
 	}
 
-	void PlainRtpTransport::SendRtcpPacket(RTC::RTCP::Packet* packet)
+	void PlainTransport::SendRtcpPacket(RTC::RTCP::Packet* packet)
 	{
 		MS_TRACE();
 
@@ -463,6 +663,9 @@ namespace RTC
 
 		const uint8_t* data = packet->GetData();
 		size_t len          = packet->GetSize();
+
+		if (HasSrtp() && !this->srtpSendSession->EncryptRtcp(&data, &len))
+			return;
 
 		if (this->rtcpMux)
 			this->tuple->Send(data, len);
@@ -473,7 +676,7 @@ namespace RTC
 		RTC::Transport::DataSent(len);
 	}
 
-	void PlainRtpTransport::SendRtcpCompoundPacket(RTC::RTCP::CompoundPacket* packet)
+	void PlainTransport::SendRtcpCompoundPacket(RTC::RTCP::CompoundPacket* packet)
 	{
 		MS_TRACE();
 
@@ -482,6 +685,9 @@ namespace RTC
 
 		const uint8_t* data = packet->GetData();
 		size_t len          = packet->GetSize();
+
+		if (HasSrtp() && !this->srtpSendSession->EncryptRtcp(&data, &len))
+			return;
 
 		if (this->rtcpMux)
 			this->tuple->Send(data, len);
@@ -492,7 +698,7 @@ namespace RTC
 		RTC::Transport::DataSent(len);
 	}
 
-	void PlainRtpTransport::SendSctpData(const uint8_t* data, size_t len)
+	void PlainTransport::SendSctpData(const uint8_t* data, size_t len)
 	{
 		MS_TRACE();
 
@@ -505,8 +711,7 @@ namespace RTC
 		RTC::Transport::DataSent(len);
 	}
 
-	inline void PlainRtpTransport::OnPacketReceived(
-	  RTC::TransportTuple* tuple, const uint8_t* data, size_t len)
+	inline void PlainTransport::OnPacketReceived(RTC::TransportTuple* tuple, const uint8_t* data, size_t len)
 	{
 		MS_TRACE();
 
@@ -534,10 +739,37 @@ namespace RTC
 		}
 	}
 
-	inline void PlainRtpTransport::OnRtpDataReceived(
+	inline void PlainTransport::OnRtpDataReceived(
 	  RTC::TransportTuple* tuple, const uint8_t* data, size_t len)
 	{
 		MS_TRACE();
+
+		if (HasSrtp() && !IsSrtpReady())
+			return;
+
+		// Decrypt the SRTP packet.
+		if (HasSrtp() && !this->srtpRecvSession->DecryptSrtp(const_cast<uint8_t*>(data), &len))
+		{
+			RTC::RtpPacket* packet = RTC::RtpPacket::Parse(data, len);
+
+			if (!packet)
+			{
+				MS_WARN_TAG(srtp, "DecryptSrtp() failed due to an invalid RTP packet");
+			}
+			else
+			{
+				MS_WARN_TAG(
+				  srtp,
+				  "DecryptSrtp() failed [ssrc:%" PRIu32 ", payloadType:%" PRIu8 ", seq:%" PRIu16 "]",
+				  packet->GetSsrc(),
+				  packet->GetPayloadType(),
+				  packet->GetSequenceNumber());
+
+				delete packet;
+			}
+
+			return;
+		}
 
 		// If multiSource is set allow it without any checking.
 		if (this->multiSource)
@@ -566,7 +798,16 @@ namespace RTC
 
 			// If not yet connected do it now.
 			if (!wasConnected)
+			{
+				// Notify the Node PlainTransport.
+				json data = json::object();
+
+				this->tuple->FillJson(data["tuple"]);
+
+				Channel::Notifier::Emit(this->id, "tuple", data);
+
 				RTC::Transport::Connected();
+			}
 		}
 		// Otherwise, if RTP tuple is set, verify that it matches the origin
 		// of the packet.
@@ -590,10 +831,19 @@ namespace RTC
 		RTC::Transport::ReceiveRtpPacket(packet);
 	}
 
-	inline void PlainRtpTransport::OnRtcpDataReceived(
+	inline void PlainTransport::OnRtcpDataReceived(
 	  RTC::TransportTuple* tuple, const uint8_t* data, size_t len)
 	{
 		MS_TRACE();
+
+		if (HasSrtp() && !IsSrtpReady())
+			return;
+
+		// Decrypt the SRTCP packet.
+		if (HasSrtp() && !this->srtpRecvSession->DecryptSrtcp(const_cast<uint8_t*>(data), &len))
+		{
+			return;
+		}
 
 		// If multiSource is set allow it without any checking.
 		if (this->multiSource)
@@ -622,7 +872,16 @@ namespace RTC
 
 			// If not yet connected do it now.
 			if (!wasConnected)
+			{
+				// Notify the Node PlainTransport.
+				json data = json::object();
+
+				this->tuple->FillJson(data["tuple"]);
+
+				Channel::Notifier::Emit(this->id, "tuple", data);
+
 				RTC::Transport::Connected();
+			}
 		}
 		// Otherwise, if RTCP-mux is unset and RTCP tuple is unset, set it if we
 		// are in comedia mode.
@@ -641,6 +900,13 @@ namespace RTC
 
 			if (!this->listenIp.announcedIp.empty())
 				this->rtcpTuple->SetLocalAnnouncedIp(this->listenIp.announcedIp);
+
+			// Notify the Node PlainTransport.
+			json data = json::object();
+
+			this->rtcpTuple->FillJson(data["rtcpTuple"]);
+
+			Channel::Notifier::Emit(this->id, "rtcpTuple", data);
 		}
 		// If RTCP-mux verify that the packet's tuple matches our RTP tuple.
 		else if (this->rtcpMux && !this->tuple->Compare(tuple))
@@ -670,7 +936,7 @@ namespace RTC
 		RTC::Transport::ReceiveRtcpPacket(packet);
 	}
 
-	inline void PlainRtpTransport::OnSctpDataReceived(
+	inline void PlainTransport::OnSctpDataReceived(
 	  RTC::TransportTuple* tuple, const uint8_t* data, size_t len)
 	{
 		MS_TRACE();
@@ -704,7 +970,16 @@ namespace RTC
 
 			// If not yet connected do it now.
 			if (!wasConnected)
+			{
+				// Notify the Node PlainTransport.
+				json data = json::object();
+
+				this->tuple->FillJson(data["tuple"]);
+
+				Channel::Notifier::Emit(this->id, "tuple", data);
+
 				RTC::Transport::Connected();
+			}
 		}
 		// Otherwise, if RTP tuple is set, verify that it matches the origin
 		// of the packet.
@@ -719,7 +994,7 @@ namespace RTC
 		RTC::Transport::ReceiveSctpData(data, len);
 	}
 
-	inline void PlainRtpTransport::OnUdpSocketPacketReceived(
+	inline void PlainTransport::OnUdpSocketPacketReceived(
 	  RTC::UdpSocket* socket, const uint8_t* data, size_t len, const struct sockaddr* remoteAddr)
 	{
 		MS_TRACE();